Split up the logic in multiple files
